requirements.txt
numpy>=1.23.5
pillow>=8.4.0
torch>=1.13.1
torchvision>=0.14.1
